commit 6

+++ schedule/src/main/java/Services/RezervareService.java
@@ -5,9 +5,9 @@ import model.Rezervare;
 import java.util.List;
 
 public interface RezervareService {
-    Rezervare findById(Long id);
+    Rezervare findById(int id);
     Rezervare add(Rezervare rezervare);
     List<Rezervare> listAll();
-    void remove(Long id);
+    void remove(int id);
     Rezervare update(Rezervare rezervare);
 }
