+++ schedule/src/main/java/Services/SalaService.java
@@ -6,9 +6,9 @@ import java.util.LinkedHashMap;
 import java.util.List;
 
 public interface SalaService{
-    Sala findById(Long id);
+    Sala findById(int id);
         Sala add(Sala sala);
         List<Sala> listAll();
-        void remove(Long id);
+        void remove(int id);
         Sala update(Sala sala);
 }
